Refactored calculation code to allow better reuse of calculation artifacts
